refactor: refactor arrange the shouldNotFilter in JWTAuthenticationFilter
