+++ src/main/java/com/bit/lot/flower/auth/common/valueobject/JWTAuthenticationShouldNotFilterAntMatcher.java
@@ -5,4 +5,7 @@ public class JWTAuthenticationShouldNotFilterAntMatcher {
   public static final String LOGIN_ANT = "/login";
   public static final String EMAIL_ANT = "/emails";
   public static final String SIGNUP_ANT = "/signup";
+    public static final String REFRESH_ANT = "/refresh";
+
+
 }
